Rename MCPARReviewSubmitPage to ReviewSubmitPage
Since MLR now uses the page too.

also rename
